+++ src/asgard_api_auditor/__init__.py
@@ -1,3 +1,3 @@
 """ASGARD API Auditor."""
 
-__version__ = "0.5.3"
+__version__ = "0.5.4"